* commit 'c29be6ea78021cb832e04b2752c3f47e8c79c422': (29 commits)
  Fix merge errors
  Fix se soluciona multitask en consulta a db
  Se genera versión para revisión
  Se resuelve conflicto con los pods
  se corrige a automaica
  Se oculta botón de finalizar orden cuando se está en status terminado
  Fix grafica
  Actualizacion ng test
  Actualiacion de grafica
  Se corrige error al procesar duplicadas
  Se modifica propiedad para actualizacion de detalles de orden de fabricacion
  Actualizacion
  Se agrega weak self
  Actualizacion de misComponentes
  environments
  Actualizacion de misComponentes
  Se agrega la propiedad destiny para traer el destino desde el servicio
  Se arregla issue con los iconos de agrupación
  Endpoints
  Cmbios localos en mis compoenentes
  ...

+++ services/sapdiapi-service/Omicron.SapDiApi.Entities/Models/CompleteDetalleFormulaModel.cs
@@ -47,7 +47,7 @@ namespace Omicron.SapDiApi.Entities.Models
         /// Gets or sets Code.
         /// </summary>
         /// <value>The code.</value>
-        public int Consumed { get; set; }
+        public decimal Consumed { get; set; }
 
         /// <summary>
         /// Gets or sets Code.
